backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
sympy==1.12
numpy==1.26.4
matplotlib==3.8.3
